refactor(platform): collapse xstate machine; GameTransitions is sole authority
Delete the mirrored XState machine (gameMachine.ts) and GameMachineService,
which had zero runtime callers and no @statelyai/inspect dependency — their
only role was a second encoding of the lifecycle kept in lock-step by a
property test. GAME_TRANSITIONS is now the single source of truth.

- fold getGameLifecycleState (+ GameRowForLifecycle, GameLifecycleState) into
  GameTransitions.ts; derive canActivate* from the computed availableEvents
- drop the GameMachineService re-export from index.ts
- remove the xstate dependency (package.json + pnpm-lock.yaml)
- add scripts/lifecycle-diagram.ts (Mermaid stateDiagram from GAME_TRANSITIONS)
  + lifecycle:diagram script to replace the lost Stately visualization
- new GameTransitions.test.ts ports the portable machine-test coverage
- GameService.ts: rename the dev-only shadow log/doc ([xstate] ->
  [lifecycle-shadow]); XSTATE_SHADOW env var kept for back-compat
