+++ packages/platform/package.json
@@ -11,8 +11,7 @@
   "dependencies": {
     "nanoid": "5.0.8",
     "random-js": "2.1.0",
-    "winston": "3.17.0",
-    "xstate": "5.20.1"
+    "winston": "3.17.0"
   },
   "devDependencies": {
     "@rollup/plugin-node-resolve": "~15.3.0",
@@ -56,7 +55,8 @@
     "build:ts": "cross-env NODE_ENV=production rollup -c",
     "dev": "cross-env NODE_ENV=development rollup -c --watch",
     "generate": "prisma generate",
-    "test": "tsx --test 'src/**/*.test.ts'"
+    "test": "tsx --test 'src/**/*.test.ts'",
+    "lifecycle:diagram": "tsx scripts/lifecycle-diagram.ts"
   },
   "engines": {
     "node": ">=24"
